Self contained test header
